api2/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
facenet-pytorch==2.5.3
pillow==10.4.0
numpy==1.26.4
torch==2.4.1
python-multipart==0.0.9
